+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2017-11-01 14:40:41 +0100 using RuboCop version 0.51.0.
+# on 2017-11-01 19:43:26 +0100 using RuboCop version 0.51.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -13,9 +13,9 @@ Metrics/AbcSize:
 # Offense count: 37
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 190
+  Max: 173
 
-# Offense count: 157
+# Offense count: 158
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
   Max: 15
