+++ lib/rubocop/config_loader.rb
@@ -16,8 +16,6 @@ module RuboCop
     AUTO_GENERATED_FILE = '.rubocop_todo.yml'.freeze
 
     class << self
-      include ConfigLoaderResolver
-
       attr_accessor :debug, :auto_gen_config, :ignore_parent_exclusion
       attr_writer :root_level # The upwards search is stopped at this level.
       attr_writer :default_configuration
@@ -37,13 +35,13 @@ module RuboCop
         hash = load_yaml_configuration(path)
 
         # Resolve requires first in case they define additional cops
-        resolve_requires(path, hash)
+        resolver.resolve_requires(path, hash)
 
         add_missing_namespaces(path, hash)
         target_ruby_version_to_f!(hash)
 
-        resolve_inheritance_from_gems(hash, hash.delete('inherit_gem'))
-        resolve_inheritance(path, hash, file)
+        resolver.resolve_inheritance_from_gems(hash, hash.delete('inherit_gem'))
+        resolver.resolve_inheritance(path, hash, file)
 
         hash.delete('inherit_from')
 
@@ -65,33 +63,7 @@ module RuboCop
       # with the addition that any value that is a hash, and occurs in both
       # arguments, will also be merged. And so on.
       def merge(base_hash, derived_hash)
-        result = base_hash.merge(derived_hash)
-        keys_appearing_in_both = base_hash.keys & derived_hash.keys
-        keys_appearing_in_both.each do |key|
-          next unless base_hash[key].is_a?(Hash)
-          result[key] = merge(base_hash[key], derived_hash[key])
-        end
-        result
-      end
-
-      def base_configs(path, inherit_from, file)
-        configs = Array(inherit_from).compact.map do |f|
-          load_file(inherited_file(path, f, file))
-        end
-
-        configs.compact
-      end
-
-      def inherited_file(path, inherit_from, file)
-        regex = URI::DEFAULT_PARSER.make_regexp(%w[http https])
-        if inherit_from =~ /\A#{regex}\z/
-          RemoteConfig.new(inherit_from, File.dirname(path))
-        elsif file.is_a?(RemoteConfig)
-          file.inherit_from_remote(inherit_from, path)
-        else
-          print 'Inheriting ' if debug?
-          File.expand_path(inherit_from, File.dirname(path))
-        end
+        resolver.merge(base_hash, derived_hash)
       end
 
       # Returns the path of .rubocop.yml searching upwards in the
@@ -136,22 +108,7 @@ module RuboCop
       # so that only cops explicitly disabled in user configuration are
       # disabled.
       def merge_with_default(config, config_file)
-        default_configuration = self.default_configuration
-
-        disabled_by_default = config.for_all_cops['DisabledByDefault']
-        enabled_by_default = config.for_all_cops['EnabledByDefault']
-
-        if disabled_by_default || enabled_by_default
-          default_configuration = transform(default_configuration) do |params|
-            params.merge('Enabled' => !disabled_by_default)
-          end
-        end
-
-        if disabled_by_default
-          config = handle_disabled_by_default(config, default_configuration)
-        end
-
-        Config.new(merge(default_configuration, config), config_file)
+        resolver.merge_with_default(config, config_file)
       end
 
       def target_ruby_version_to_f!(hash)
@@ -182,34 +139,8 @@ module RuboCop
 
       private
 
-      def handle_disabled_by_default(config, new_default_configuration)
-        department_config = config.to_hash.reject { |cop| cop.include?('/') }
-        department_config.each do |dept, dept_params|
-          # Rails is always disabled by default and the department's Enabled
-          # flag works like the --rails command line option, which is that when
-          # AllCops:DisabledByDefault is true, each Rails cop must still be
-          # explicitly mentioned in user configuration in order to be enabled.
-          next if dept == 'Rails'
-
-          next unless dept_params['Enabled']
-
-          new_default_configuration.each do |cop, params|
-            next unless cop.start_with?(dept + '/')
-
-            # Retain original default configuration for cops in the department.
-            params['Enabled'] = default_configuration[cop]['Enabled']
-          end
-        end
-
-        transform(config) do |params|
-          { 'Enabled' => true }.merge(params) # Set true if not set.
-        end
-      end
-
-      # Returns a new hash where the parameters of the given config hash have
-      # been replaced by parameters returned by the given block.
-      def transform(config)
-        Hash[config.map { |cop, params| [cop, yield(params)] }]
+      def resolver
+        @resolver ||= ConfigLoaderResolver.new
       end
 
       def load_yaml_configuration(absolute_path)
@@ -246,14 +177,6 @@ module RuboCop
         else
           YAML.load(yaml_code, filename) # rubocop:disable Security/YAMLLoad
         end
-      end
-
-      def gem_config_path(gem_name, relative_config_path)
-        spec = Gem::Specification.find_by_name(gem_name)
-        return File.join(spec.gem_dir, relative_config_path)
-      rescue Gem::LoadError => e
-        raise Gem::LoadError,
-              "Unable to find gem #{gem_name}; is the gem installed? #{e}"
       end
 
       def config_files_in_path(target)
